Fix parsing when using large header buffers
If headers are large, nginx may switch to a
large header buffer (8K for example).
If the POST request fits into this buffer,
it will be contained into the preread size
and ngx_http_request_body_filter will read
everything leaving nothing more to do.
So we just cleanup and go on.